requirements.txt
Django==4.0.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
mysqlclient<=2.1.0  # Only needed if using MySQL